Refactor code to make it more modular
The code is now more modular, with each operation being a function. I also created an array of some of the most popular tech companies and their ticker symbols. I parsed through this to get the stock and article information for the first 3 pages of each query, which I then put into two different files. One with only the stock change and sentiment, and one with the text, change, and sentiment.